Module Ppx_bench_lib

module Benchmark_accumulator : sig ... end

The point of Benchmark_accumulator is to provide a global place where inline benchmarking macros can register themselves. Once registered here, the benchmarks are retrieved and analyzed using Core_bench.

module Export : sig ... end